+++ CHANGELOG.md
@@ -8,6 +8,8 @@ Each release is tagged in Git and on the Docker repository
 
 ## unreleased
 
+*   introduce a configuration file `/etc/moonfire-nvr.json`; you will need
+    to create one when upgrading.
 *   bump minimum Rust version from 1.53 to 1.56.
 *   fix [#187](https://github.com/scottlamb/moonfire-nvr/issues/187):
     incompatibility with cameras that (incorrectly) omit the SDP origin line.